+++ CHANGELOG.md
@@ -5,7 +5,8 @@
   * Issues Fixed:
     * `faspex5`: Fix public link auth for Faspex 5.0.13.
     * `aoc`: Fix some admin operations requiring a user's home for Files.
-
+  * Breaking Changes:
+    * `ats`: Removed option `params`. Use positinal parameter for creation, and `query` for list.
 * 4.24.1
 
   * Issues Fixed: